On the one hand, this is a completely pointless commit.  All I did was switch from closestfare to mixedmode in overrides.ini.
On the other hand, it's the farthest thing from it.  Check it out!  mixedmode
works great!  And it's firing on all (or almost all anyway) cylinders!  There
are no pickups from Yellow-0, but that may be because it was a short run, or
it may be the same bug that closestfare exhibits.

C:\Source\hg\unified>agents_driver.py| grep -E "got Fare|waitingFares"
.. Yellow-4 got Fare -2
.. Yellow-3 got Fare -3
.. Yellow-3 got Fare -4
.. Yellow-2 got Fare 1
.. Yellow-4 got Fare 0
.. Yellow-1 got Fare 2
.. Yellow-4 got Fare -1
.. Yellow-3 got Fare 3
.. Yellow-1 got Fare 7
.. Yellow-4 got Fare 9
.. Yellow-2 got Fare 10
.. Yellow-3 got Fare 6
.. Yellow-1 got Fare 14
.. Yellow-3 got Fare 12
.. Yellow-1 got Fare 4
.. Yellow-3 got Fare 13
.. Yellow-1 got Fare 8
.. Yellow-1 got Fare 5
.. Yellow-1 got Fare 27
.. Yellow-1 got Fare 17
.. Yellow-1 got Fare 15
waitingFares [11, 16, 18, 19, 20, 21, 22, 23, 24, 25, 26, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39]