I need to fix the problem with graphs having disconnected regions, like this:
   *------*   *
               \
                \
                 *

To that end, I added module networkx.component and a little test method
(get_connected()).  So far so good!  I may be able to just use the largest
component, ie, the first one in the list of connected_components, and add the
rest to a dead list, but if that doesn't work, I may have to remove them from
the database and then regenerate the graph (hopefully it won't come to that).
Not sure yet.


From the NetworkX documentation:

connected_components(G)

Return a list of lists of nodes in each connected component of G.

The list is ordered from largest connected component to smallest. For undirected graphs only.