Big commit.  condit_thesis.lyx is now a container with the front and back matter, but the chapters are broken out into separate (unnumbered!) sections.
One thing I've not figured out yet is how to work with them like this.  I
think for now I'm just going to keep them separate, and look at individual
chapters most of the time.  Every so often it would be a good idea to
reassemble the whole thing and view it in its entirety.